third_party/blink/web_tests/storage/indexeddb/mozilla/key-requirements-put-no-key-expected.txt

Test IndexedDB's behavior puting without key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "key-requirements-put-no-key.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('bar');
Expecting exception from objectStore.put({});
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: Failed to execute 'put' on 'IDBObjectStore': The object store uses out-of-line keys and has no key generator and the key parameter was not provided.
PASS successfullyParsed is true

TEST COMPLETE
